// lib/api/generated/models/meetingReviewResponseDataItemsItemKind.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Heymoa API
 * Heymoa 서버 REST API
 * OpenAPI spec version: 1.0.0
 */

/**
 * 항목 종류
 */
export type MeetingReviewResponseDataItemsItemKind =
  (typeof MeetingReviewResponseDataItemsItemKind)[keyof typeof MeetingReviewResponseDataItemsItemKind];

export const MeetingReviewResponseDataItemsItemKind = {
  AGENDA: "AGENDA",
  DECISION: "DECISION",
  ACTION_ITEM: "ACTION_ITEM",
  ISSUE: "ISSUE",
  QUESTION: "QUESTION",
  STATUS_REPORT: "STATUS_REPORT",
  INSIGHT: "INSIGHT",
} as const;